+++ frontend/src/pages/SignUpPage.tsx
@@ -1,5 +1,3 @@
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
   Card,
@@ -12,130 +10,10 @@ import Footer from "@/components/layout/Footer";
 import Header from "@/components/layout/Header";
 import PageContainer from "@/components/layout/PageContainer";
 import Logo from "@/components/ui/Logo";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { z } from "zod";
-
-const formSchema = z
-  .object({
-    fullName: z
-      .string({
-        required_error: "Name is required",
-        invalid_type_error:
-          "full name should be between 4 and 50 characters long. Try again please",
-      })
-      .min(4)
-      .max(50),
-    email: z
-      .string({
-        required_error: "email is required",
-        invalid_type_error: "email is not a valid email. Try again please",
-      })
-      .email(),
-    password: z
-      .string({
-        required_error: "password is required",
-        invalid_type_error:
-          "password should be between 8 and 20 characters long. Try again please",
-      })
-      .min(8)
-      .max(20),
-    passwordConfirm: z
-      .string({
-        required_error: "confirm password is required",
-        invalid_type_error:
-          "confirm password should be between 8 and 20 characters long. Try again please",
-      })
-      .min(8)
-      .max(20),
-  })
-  .refine((data) => data.password === data.passwordConfirm, {
-    message: "Passwords must match",
-    path: ["passwordConfirm"], // This specifies where the error should be attached in case of mismatch
-  });
-
-const signUpFormFields = [
-  {
-    name: "fullName",
-    label: "full name",
-    placeHolder: "Your Name",
-    type: "text",
-  },
-  {
-    name: "email",
-    label: "Email",
-    placeHolder: "example@email.com",
-    type: "email",
-  },
-  {
-    name: "password",
-    label: "password",
-    placeHolder: "********",
-    type: "password",
-  },
-  {
-    name: "passwordConfirm",
-    label: "Password confirm",
-    placeHolder: "********",
-    type: "password",
-  },
-] as const;
-const loginFormFields = [
-  {
-    name: "email",
-    label: "Email",
-    placeHolder: "example@email.com",
-    type: "email",
-  },
-  {
-    name: "password",
-    label: "password",
-    placeHolder: "********",
-    type: "password",
-  },
-] as const;
+import LoginForm from "@/components/form/LoginForm";
+import SignUpForm from "@/components/form/SignUpForm";
 
 export default function SignUpPage() {
-  // 1. Define your form.
-  const signUpForm = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      fullName: "",
-      email: "",
-      password: "",
-      passwordConfirm: "",
-    },
-  });
-  const loginForm = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-  });
-
-  function handleSignup(values: z.infer<typeof formSchema>) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
-    console.log(values);
-    signUpForm.reset();
-  }
-
-  function handleLogin(values: z.infer<typeof formSchema>) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
-    console.log(values);
-    loginForm.reset();
-  }
-
   return (
     <PageContainer>
       <div className="flex min-h-dvh flex-col">
@@ -161,6 +39,7 @@ export default function SignUpPage() {
                       disableLink
                     />
                   </CardTitle>
+
                   <CardDescription className="text-center">
                     <span className="mb-4 mt-2 block text-lg font-bold">
                       Sign up
@@ -170,40 +49,9 @@ export default function SignUpPage() {
                     </span>
                   </CardDescription>
                 </CardHeader>
-                <CardContent>
-                  <Form {...signUpForm}>
-                    <form
-                      onSubmit={signUpForm.handleSubmit(handleSignup)}
-                      className="space-y-8"
-                    >
-                      {signUpFormFields.map(
-                        ({ name, label, placeHolder, type }) => (
-                          <FormField
-                            key={name + label + placeHolder}
-                            control={signUpForm.control}
-                            name={name}
-                            render={({ field }) => (
-                              <FormItem>
-                                <FormLabel>{label}</FormLabel>
-                                <FormControl>
-                                  <Input
-                                    type={type}
-                                    placeholder={placeHolder}
-                                    {...field}
-                                  />
-                                </FormControl>
-                                <FormMessage />
-                              </FormItem>
-                            )}
-                          />
-                        ),
-                      )}
 
-                      <Button type="submit" className="w-full">
-                        Submit
-                      </Button>
-                    </form>
-                  </Form>
+                <CardContent>
+                  <SignUpForm />
                 </CardContent>
               </Card>
             </TabsContent>
@@ -218,6 +66,7 @@ export default function SignUpPage() {
                       className="mb-2 scale-125 p-1"
                     />
                   </CardTitle>
+
                   <CardDescription className="text-center">
                     <span className="mb-4 mt-2 block text-lg font-bold">
                       Login
@@ -227,40 +76,9 @@ export default function SignUpPage() {
                     </span>
                   </CardDescription>
                 </CardHeader>
-                <CardContent>
-                  <Form {...loginForm}>
-                    <form
-                      onSubmit={loginForm.handleSubmit(handleLogin)}
-                      className="space-y-8"
-                    >
-                      {loginFormFields.map(
-                        ({ name, label, placeHolder, type }) => (
-                          <FormField
-                            key={name + label + placeHolder}
-                            control={loginForm.control}
-                            name={name}
-                            render={({ field }) => (
-                              <FormItem>
-                                <FormLabel>{label}</FormLabel>
-                                <FormControl>
-                                  <Input
-                                    type={type}
-                                    placeholder={placeHolder}
-                                    {...field}
-                                  />
-                                </FormControl>
-                                <FormMessage />
-                              </FormItem>
-                            )}
-                          />
-                        ),
-                      )}
 
-                      <Button type="submit" className="w-full">
-                        Submit
-                      </Button>
-                    </form>
-                  </Form>
+                <CardContent>
+                  <LoginForm />
                 </CardContent>
               </Card>
             </TabsContent>
